requirements.txt
opencv-python
mediapipe
Pillow
customtkinter
numpy  
pandas